refactor(web): split chart into helpers, styles, and component files
- Extract chartHelpers.ts with types, formatters, API calls
- Extract Sparkline.module.scss for chart-specific styles
- TokenClient.module.scss back under 500 lines
- MarketCapSparkline.tsx back under 500 lines
- Fix circular type imports
